Restructured the notes Pane & Formik components
